chore: worked on some repo functions
